requirements.txt
streamlit
jira
python-dotenv
langchain
anthropic
pydantic